requirements.txt
Django~=5.2.3
djangorestframework
markdown
django-filter
Pillow
django-cors-headers
requests
dotenv
openai